feat: add label cmd
